requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings
typing-extensions>=4.0.0

# Async web framework and server
fastapi>=0.100.1
uvicorn[standard]

# Message broker client
pika>=1.3.2
aio-pika

# Async HTTP client
aiohttp>=3.9.1
aiofiles>=23.2.0

# Server-Sent Events for streaming
sse-starlette>=1.6.5

# Prometheus monitoring
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# Environment and Utils
python-dotenv==1.0.0
requests>=2.31.0
orjson>=3.9.10

# MCP Specific (The "Superpowers")
mcp
langchain-mcp-adapters
langchain-openai
requests
ollama

# Web UI
streamlit>=1.40.0

